notify end

+++ www/delete.php
@@ -14,30 +14,26 @@ if ($_SERVER['REQUEST_METHOD'] === "POST") {
 
 	// удаление изображения фильма
 	$film = get_film($_GET['id']);
-	if (is_file(ROOT . 'data/films/' . $film['photo'])) {
-		unlink(ROOT . 'data/films/' . $film['photo']);
-	}
-	if (is_file(ROOT . 'data/films/min/' . $film['photo'])) {
-		unlink(ROOT . 'data/films/min/' . $film['photo']);
-	}
-	if (is_file(ROOT . 'data/films/big/' . $film['photo'])) {
-		unlink(ROOT . 'data/films/big/' . $film['photo']);
-	}
+	if (is_file(ROOT . 'data/films/' . $film['photo'])) unlink(ROOT . 'data/films/' . $film['photo']);
 
+	if (is_file(ROOT . 'data/films/min/' . $film['photo'])) unlink(ROOT . 'data/films/min/' . $film['photo']);
+
+	if (is_file(ROOT . 'data/films/big/' . $film['photo'])) unlink(ROOT . 'data/films/big/' . $film['photo']);
 
 	// удаление фильма из БД
 	$result = delete_film($_GET['id']);
 
 	if ($result) {
+		$_SESSION['success'][] = "Фильм успешно удалён";
 		header('Location: ' . HOST);
 		exit;
 	} else {
-		$error = "Ошибка при удалении фильма";
+		$_SESSION['errors'][] = "Ошибка при удалении фильма";
 	}
 }
 
 $film = get_film($_GET['id']);
-
+if ($film) $page_title = $film['title'];
 
 
 include(ROOT . 'templates/head.tpl');
@@ -47,25 +43,15 @@ include(ROOT . 'templates/header.tpl');
 <main class="main">
 	<div class="container">
 
-		<?php echo isset($error) ? $error : null ?>
-
-		<?php if (isset($error)): ?>
-			<div class="alert-wrapper">
-				<div class="alert alert--error"><?= $error ?></div>
-			</div>
-		<?php endif; ?>
+		<?php include(ROOT . "templates/errors.tpl"); ?>
 
 		<?php if ($film): ?>
-
 			<h1 class="title-1">Удалить фильм?</h1>
 			<?php include(ROOT . 'templates/form-delete.tpl'); ?>
-
 		<?php else: ?>
-
 			<div class="alert-wrapper">
-				<div class="alert alert--warning">Такого фильма не существует</div>
+				<?php echo notify("Такого фильма не существует") ?>
 			</div>
-
 		<?php endif; ?>
